20230629H

+++ main.ps1
@@ -4,14 +4,42 @@ Get-Alias -Scope 'Local' -ErrorAction 'SilentlyContinue' |
 	Remove-Alias -Scope 'Local' -Force -ErrorAction 'SilentlyContinue'
 Import-Module -Name 'hugoalh.GitHubActionsToolkit' -Scope 'Local'
 Test-GitHubActionsEnvironment -Mandatory
-Write-Host -Object 'Import inputs.'
-[String]$SessionID = New-Guid |
-	Select-Object -ExpandProperty 'Guid'
+Write-Host -Object 'Initialize.'
+[String]$JobIdPrefix = (
+	New-Guid |
+		Select-Object -ExpandProperty 'Guid'
+).ToUpper() -ireplace '-', ''
 [Boolean]$OsLinux = $Env:RUNNER_OS -ieq 'Linux'
 [Boolean]$OsMac = $Env:RUNNER_OS -ieq 'MacOS'
 [Boolean]$OsWindows = $Env:RUNNER_OS -ieq 'Windows'
 [String]$OsPathType = "Path$($Env:RUNNER_OS)"
+Try {
+	$ProgramAPT = Get-Command -Name 'apt-get' -CommandType 'Application' -ErrorAction 'Stop'
+}
+Catch {}
+Try {
+	$ProgramChocolatey = Get-Command -Name 'choco' -CommandType 'Application' -ErrorAction 'Stop'
+}
+Catch {}
+Try {
+	$ProgramDocker = Get-Command -Name 'docker' -CommandType 'Application' -ErrorAction 'Stop'
+}
+Catch {}
+Try {
+	$ProgramHomebrew = Get-Command -Name 'brew' -CommandType 'Application' -ErrorAction 'Stop'
+}
+Catch {}
+Try {
+	$ProgramNPM = Get-Command -Name 'npm' -CommandType 'Application' -ErrorAction 'Stop'
+}
+Catch {}
+Try {
+	$ProgramPipx = Get-Command -Name 'pipx' -CommandType 'Application' -ErrorAction 'Stop'
+}
+Catch {}
+Write-Host -Object 'Import inputs.'
 [RegEx]$InputListDelimiter = Get-GitHubActionsInput -Name 'input_listdelimiter' -Mandatory -EmptyStringAsNull
+[Boolean]$OperationAsync = [Boolean]::Parse((Get-GitHubActionsInput -Name 'operate_async' -Mandatory -EmptyStringAsNull))
 [AllowEmptyCollection()][RegEx[]]$RemoveGeneralInclude = ((
 	((Get-GitHubActionsInput -Name 'general_include' -EmptyStringAsNull) ?? '') -isplit $InputListDelimiter |
 		Where-Object -FilterScript { $_.Length -gt 0 }
@@ -38,7 +66,6 @@ Write-Host -Object 'Import inputs.'
 [Boolean]$RemoveHomebrewCache = [Boolean]::Parse((Get-GitHubActionsInput -Name 'homebrewcache' -Mandatory -EmptyStringAsNull))
 [Boolean]$RemoveNpmCache = [Boolean]::Parse((Get-GitHubActionsInput -Name 'npmcache' -Mandatory -EmptyStringAsNull))
 [Boolean]$RemoveLinuxSwap = [Boolean]::Parse((Get-GitHubActionsInput -Name 'swap' -Mandatory -EmptyStringAsNull))
-[Boolean]$OperationAsync = [Boolean]::Parse((Get-GitHubActionsInput -Name 'op_async' -EmptyStringAsNull))
 Function Get-DiskSpace {
 	[CmdletBinding()]
 	[OutputType([String])]
@@ -72,11 +99,11 @@ Function Test-StringMatchRegEx {
 [String]$DiskSpaceBefore = Get-DiskSpace
 $Script:ErrorActionPreference = 'Continue'
 <# Docker Image. #>
-If ($OsLinux -or $OsWindows) {
+If ($Null -ine $ProgramDocker) {
 	If ($OperationAsync) {
 		Write-Host -Object 'Prune/Remove Docker images.'
-		$Null = Start-Job -Name "$SessionID-Docker" -ScriptBlock {
-			If ($Using:RemoveDockerImageInclude.Count -gt 0) {
+		$Null = Start-Job -Name "$JobIdPrefix/Docker" -ScriptBlock {
+			If (($Using:RemoveDockerImageInclude).Count -gt 0) {
 				[PSCustomObject[]]$DockerImageList = (
 					docker image ls --all --format '{{json .}}' |
 						Join-String -Separator ',' -OutputPrefix '[' -OutputSuffix ']' |
@@ -124,11 +151,11 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 	ForEach ($Item In (
 		Import-Csv -LiteralPath (Join-Path -Path $PSScriptRoot -ChildPath 'list.tsv') -Delimiter "`t" -Encoding 'UTF8NoBOM' -ErrorAction 'Continue' |
 			Where-Object -FilterScript {
-				($OsLinux -and $_.APT.Length -gt 0) -or
-				($OsWindows -and $_.Chocolatey.Length -gt 0) -or
-				($OsMac -and $_.Homebrew.Length -gt 0) -or
-				$_.NPM.Length -gt 0 -or
-				(($OsLinux -or $OsMac) -and $_.Pipx.Length -gt 0) -or
+				($Null -ine $ProgramAPT -and $_.APT.Length -gt 0) -or
+				($Null -ine $ProgramChocolatey -and $_.Chocolatey.Length -gt 0) -or
+				($Null -ine $ProgramHomebrew -and $_.Homebrew.Length -gt 0) -or
+				($Null -ine $ProgramNPM -and $_.NPM.Length -gt 0) -or
+				($Null -ine $ProgramPipx -and $_.Pipx.Length -gt 0) -or
 				$_.Env.Length -gt 0 -or
 				$_.($OsPathType).Length -gt 0
 			} |
@@ -137,13 +164,13 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 			Sort-Object -Property 'Priority' -Descending
 	)) {
 		Write-Host -Object "Remove $($Item.Description)."
-		If ($OsLinux -and $Item.APT.Length -gt 0) {
+		If ($Null -ine $ProgramAPT -and $Item.APT.Length -gt 0) {
 			If ($OperationAsync) {
-				$Null = Get-Job -Name "$SessionID-APT-*" -ErrorAction 'SilentlyContinue' |
+				$Null = Get-Job -Name "$JobIdPrefix/APT/*" -ErrorAction 'SilentlyContinue' |
 					Wait-Job
-				$Null = Start-Job -Name "$SessionID-APT-$($Item.Name)" -ScriptBlock {
+				$Null = Start-Job -Name "$JobIdPrefix/APT/$($Item.Name)" -ScriptBlock {
 					ForEach ($APT In (
-						$Using:Item.APT -isplit ';;' |
+						($Using:Item).APT -isplit ';;' |
 							Where-Object -FilterScript { $_.Length -gt 0 }
 					)) {
 						apt-get --assume-yes remove $APT |
@@ -161,13 +188,13 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 				}
 			}
 		}
-		If ($OsWindows -and $Item.Chocolatey.Length -gt 0) {
+		If ($Null -ine $ProgramChocolatey -and $Item.Chocolatey.Length -gt 0) {
 			If ($OperationAsync) {
-				$Null = Get-Job -Name "$SessionID-Chocolatey-*" -ErrorAction 'SilentlyContinue' |
+				$Null = Get-Job -Name "$JobIdPrefix/Chocolatey/*" -ErrorAction 'SilentlyContinue' |
 					Wait-Job
-				$Null = Start-Job -Name "$SessionID-Chocolatey-$($Item.Name)" -ScriptBlock {
+				$Null = Start-Job -Name "$JobIdPrefix/Chocolatey/$($Item.Name)" -ScriptBlock {
 					ForEach ($Chocolatey In (
-						$Using:Item.Chocolatey -isplit ';;' |
+						($Using:Item).Chocolatey -isplit ';;' |
 							Where-Object -FilterScript { $_.Length -gt 0 }
 					)) {
 						choco uninstall $Chocolatey --ignore-detected-reboot --yes |
@@ -185,13 +212,13 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 				}
 			}
 		}
-		If ($OsMac -and $Item.Homebrew.Length -gt 0) {
+		If ($Null -ine $ProgramHomebrew -and $Item.Homebrew.Length -gt 0) {
 			If ($OperationAsync) {
-				$Null = Get-Job -Name "$SessionID-Homebrew-*" -ErrorAction 'SilentlyContinue' |
+				$Null = Get-Job -Name "$JobIdPrefix/Homebrew/*" -ErrorAction 'SilentlyContinue' |
 					Wait-Job
-				$Null = Start-Job -Name "$SessionID-Homebrew-$($Item.Name)" -ScriptBlock {
+				$Null = Start-Job -Name "$JobIdPrefix/Homebrew/$($Item.Name)" -ScriptBlock {
 					ForEach ($Homebrew In (
-						$Using:Item.Homebrew -isplit ';;' |
+						($Using:Item).Homebrew -isplit ';;' |
 							Where-Object -FilterScript { $_.Length -gt 0 }
 					)) {
 						brew uninstall $Homebrew |
@@ -209,13 +236,13 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 				}
 			}
 		}
-		If ($Item.NPM.Length -gt 0) {
+		If ($Null -ine $ProgramNPM -and $Item.NPM.Length -gt 0) {
 			If ($OperationAsync) {
-				$Null = Get-Job -Name "$SessionID-NPM-*" -ErrorAction 'SilentlyContinue' |
+				$Null = Get-Job -Name "$JobIdPrefix/NPM/*" -ErrorAction 'SilentlyContinue' |
 					Wait-Job
-				$Null = Start-Job -Name "$SessionID-NPM-$($Item.Name)" -ScriptBlock {
+				$Null = Start-Job -Name "$JobIdPrefix/NPM/$($Item.Name)" -ScriptBlock {
 					ForEach ($NPM In (
-						$Using:Item.NPM -isplit ';;' |
+						($Using:Item).NPM -isplit ';;' |
 							Where-Object -FilterScript { $_.Length -gt 0 }
 					)) {
 						npm --global uninstall $NPM |
@@ -233,13 +260,13 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 				}
 			}
 		}
-		If (($OsLinux -or $OsMac) -and $Item.Pipx.Length -gt 0) {
+		If ($Null -ine $ProgramPipx -and $Item.Pipx.Length -gt 0) {
 			If ($OperationAsync) {
-				$Null = Get-Job -Name "$SessionID-Pipx-*" -ErrorAction 'SilentlyContinue' |
+				$Null = Get-Job -Name "$JobIdPrefix/Pipx/*" -ErrorAction 'SilentlyContinue' |
 					Wait-Job
-				$Null = Start-Job -Name "$SessionID-Pipx-$($Item.Name)" -ScriptBlock {
+				$Null = Start-Job -Name "$JobIdPrefix/Pipx/$($Item.Name)" -ScriptBlock {
 					ForEach ($Pipx In (
-						$Using:Item.Pipx -isplit ';;' |
+						($Using:Item).Pipx -isplit ';;' |
 							Where-Object -FilterScript { $_.Length -gt 0 }
 					)) {
 						pipx uninstall $Pipx |
@@ -257,11 +284,14 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 				}
 			}
 		}
-		If ($Item.Env.Length -gt 0) {
-			If ($OperationAsync) {
-				$Null = Start-Job -Name "$SessionID-Env-$($Item.Name)" -ScriptBlock {
+		If ($OperationAsync -and (
+			$Item.Env.Length -gt 0 -or
+			$Item.($OsPathType).Length -gt 0
+		)) {
+			$Null = Start-Job -Name "$JobIdPrefix/FS/$($Item.Name)" -ScriptBlock {
+				If (($Using:Item).Env.Length -gt 0) {
 					ForEach ($ItemEnv In (
-						$Using:Item.Env -isplit ';;' |
+						($Using:Item).Env -isplit ';;' |
 							Where-Object -FilterScript { $_.Length -gt 0 }
 					)) {
 						[String]$ItemEnvValue = Get-Content -LiteralPath "Env:\$ItemEnv" -ErrorAction 'SilentlyContinue'
@@ -273,27 +303,7 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 						}
 					}
 				}
-			}
-			Else {
-				ForEach ($ItemEnv In (
-					$Item.Env -isplit ';;' |
-						Where-Object -FilterScript { $_.Length -gt 0 }
-				)) {
-					[String]$ItemEnvValue = Get-Content -LiteralPath "Env:\$ItemEnv" -ErrorAction 'SilentlyContinue'
-					If ($ItemEnvValue.Length -gt 0 -and (Test-Path -LiteralPath $ItemEnvValue)) {
-						Get-ChildItem -LiteralPath $ItemEnvValue -Force -ErrorAction 'Continue' |
-							ForEach-Object -Process {
-								Remove-Item -LiteralPath $_.FullName -Recurse -Force -Confirm:$False -ErrorAction 'Continue'
-							}
-					}
-				}
-			}
-		}
-		If ($Item.($OsPathType).Length -gt 0) {
-			If ($OperationAsync) {
-				$Null = Get-Job -Name "$SessionID-Env-$($Item.Name)" -ErrorAction 'SilentlyContinue' |
-					Wait-Job
-				$Null = Start-Job -Name "$SessionID-Path-$($Item.Name)" -ScriptBlock {
+				If (($Using:Item).($Using:OsPathType).Length -gt 0) {
 					ForEach ($ItemPath In (
 						($Using:Item).($Using:OsPathType) -isplit ';;' |
 							Where-Object -FilterScript { $_.Length -gt 0 }
@@ -308,7 +318,23 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 					}
 				}
 			}
-			Else {
+		}
+		Else {
+			If ($Item.Env.Length -gt 0) {
+				ForEach ($ItemEnv In (
+					$Item.Env -isplit ';;' |
+						Where-Object -FilterScript { $_.Length -gt 0 }
+				)) {
+					[String]$ItemEnvValue = Get-Content -LiteralPath "Env:\$ItemEnv" -ErrorAction 'SilentlyContinue'
+					If ($ItemEnvValue.Length -gt 0 -and (Test-Path -LiteralPath $ItemEnvValue)) {
+						Get-ChildItem -LiteralPath $ItemEnvValue -Force -ErrorAction 'Continue' |
+							ForEach-Object -Process {
+								Remove-Item -LiteralPath $_.FullName -Recurse -Force -Confirm:$False -ErrorAction 'Continue'
+							}
+					}
+				}
+			}
+			If ($Item.($OsPathType).Length -gt 0) {
 				ForEach ($ItemPath In (
 					$Item.($OsPathType) -isplit ';;' |
 						Where-Object -FilterScript { $_.Length -gt 0 }
@@ -325,12 +351,12 @@ If ($RemoveGeneralInclude.Count -gt 0) {
 		}
 	}
 }
-$Null = Get-Job -Name "$SessionID-*" -ErrorAction 'SilentlyContinue' |
+$Null = Get-Job -Name "$JobIdPrefix/*" -ErrorAction 'SilentlyContinue' |
 	Wait-Job
-If ($OsLinux -and $RemoveAptCache) {
+If ($Null -ine $ProgramAPT -and $RemoveAptCache) {
 	Write-Host -Object 'Remove APT cache.'
 	If ($OperationAsync) {
-		$Null = Start-Job -Name "$SessionID-APTCache" -ScriptBlock {
+		$Null = Start-Job -Name "$JobIdPrefix/APTCache" -ScriptBlock {
 			apt-get --assume-yes autoremove |
 				Write-GitHubActionsDebug
 			apt-get --assume-yes clean |
@@ -344,10 +370,10 @@ If ($OsLinux -and $RemoveAptCache) {
 			Write-GitHubActionsDebug
 	}
 }
-If ($OsMac -and $RemoveHomebrewCache) {
+If ($Null -ine $ProgramHomebrew -and $RemoveHomebrewCache) {
 	Write-Host -Object 'Remove Homebrew cache.'
 	If ($OperationAsync) {
-		$Null = Start-Job -Name "$SessionID-HomebrewCache" -ScriptBlock {
+		$Null = Start-Job -Name "$JobIdPrefix/HomebrewCache" -ScriptBlock {
 			brew autoremove |
 				Write-GitHubActionsDebug
 		}
@@ -357,10 +383,10 @@ If ($OsMac -and $RemoveHomebrewCache) {
 			Write-GitHubActionsDebug
 	}
 }
-If ($RemoveNpmCache) {
+If ($Null -ine $ProgramNPM -and $RemoveNpmCache) {
 	Write-Host -Object 'Remove NPM cache.'
 	If ($OperationAsync) {
-		$Null = Start-Job -Name "$SessionID-NPMCache" -ScriptBlock {
+		$Null = Start-Job -Name "$JobIdPrefix/NPMCache" -ScriptBlock {
 			npm cache clean --force |
 				Write-GitHubActionsDebug
 		}
@@ -370,27 +396,16 @@ If ($RemoveNpmCache) {
 			Write-GitHubActionsDebug
 	}
 }
-$Null = Get-Job -Name "$SessionID-*" -ErrorAction 'SilentlyContinue' |
+$Null = Get-Job -Name "$JobIdPrefix/*" -ErrorAction 'SilentlyContinue' |
 	Wait-Job
 If ($OsLinux -and $RemoveLinuxSwap) {
 	Write-Host -Object 'Remove Linux swap space.'
-	If ($OperationAsync) {
-		$Null = Start-Job -Name "$SessionID-Swap" -ScriptBlock {
-			swapoff -a |
-				Write-GitHubActionsDebug
-			Remove-Item -LiteralPath '/mnt/swapfile' -Recurse -Force -Confirm:$False -ErrorAction 'Continue'
-		}
-	}
-	Else {
-		swapoff -a |
-			Write-GitHubActionsDebug
-		Remove-Item -LiteralPath '/mnt/swapfile' -Recurse -Force -Confirm:$False -ErrorAction 'Continue'
-	}
+	swapoff -a |
+		Write-GitHubActionsDebug
+	Remove-Item -LiteralPath '/mnt/swapfile' -Recurse -Force -Confirm:$False -ErrorAction 'Continue'
 }
-$Null = Get-Job -Name "$SessionID-*" -ErrorAction 'SilentlyContinue' |
-	Wait-Job
 If ($OperationAsync) {
-	Get-Job -Name "$SessionID-*" |
+	Get-Job -Name "$JobIdPrefix/*" |
 		Format-Table -Property @('Name', 'State') -AutoSize -Wrap |
 		Out-String -Width 120 |
 		Write-Host
